+++ src/libs/Router/path.ts
@@ -1,0 +1,7 @@
+export const path = {
+  root: '/',
+  login: '/login',
+  signup: '/signup',
+  passwordForgot: '/password-forgot',
+  newPassword: '/new-password',
+} as const;
